move back consts and functions from common. add comments
